+++ client/src/components/Header.jsx
@@ -1,4 +1,4 @@
-import './Header.Home.css';
+// import './Header.Home.css';
 function Header(){
     return (
         <header className = "home-header">
@@ -7,8 +7,8 @@ function Header(){
                     <button className = "button" href="/Home">Home</button>
                     <button className = "button" href="/Login">Login</button>
                     <button className = "button" href="/Sign up">Sign up</button>
-               <h1 classsName = "header-h1">ART OF WEAR</h1> 
-               </ul> 
+               <h1 className = "header-h1">ART OF WEAR</h1>
+               </ul>
             </nav>
         </header>
     );
