--- client/src/pages/Home.jsx
@@ -1,16 +1,0 @@
-import Post from "../components/Post";
-import Headerhome from "../components/Headerhome";
-import Footer from "../components/Footer";
-import Gallery from "../components/Gallery"
-const Home = () => {
-    return (
-        <>
-            <Headerhome />
-            <Post /> 
-            <Gallery/>
-            <Footer />
-            </>
-    );
-}
-
-export default Home;
