+++ client/src/pages/Home/Home.jsx
@@ -1,0 +1,11 @@
+import Gallery from "../../components/Gallery"
+
+const Home = () => {
+    return (
+        <>
+            <Gallery/>
+        </>
+    );
+}
+
+export default Home;
